Move Target to generator module
